+++ byteps/sparse/loop.cc
@@ -1,0 +1,12 @@
+#include "loop.h"
+#include "queue_exec.h"
+
+namespace byteps {
+namespace sparse {
+
+void runDenseReduceLoop(QueueExecLoop * queue_loop_ptr) {
+  queue_loop_ptr = QueueExecLoop::init_loop();
+}
+
+} // namespace sparse
+} // namespace byteps
