+++ byteps/sparse/sparse.cuh
@@ -1,0 +1,4 @@
+#ifndef BYTEPS_SPARSE_CUH
+#define BYTEPS_SPARSE_CUH
+
+#endif  // BYTEPS_SPARSE_CUH
